back/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[blacklist]
django-cors-headers
